move to relation tests and add more tests
testing otherclan and not working cat pregnancies
