+++ tests/test_relation_events.py
@@ -2,6 +2,14 @@ import os
 import unittest
 from unittest.mock import patch
 
+from scripts.cat.enums import CatGroup
+from scripts.clan_package.settings import (
+    load_clan_settings,
+    set_clan_setting,
+    get_clan_setting,
+)
+from scripts.game_structure import game
+
 os.environ["SDL_VIDEODRIVER"] = "dummy"
 os.environ["SDL_AUDIODRIVER"] = "dummy"
 
@@ -12,7 +20,153 @@ from scripts.events_module.relationship.pregnancy_events import Pregnancy_Events
 from scripts.events_module.relationship.romantic_events import RomanticEvents
 
 
+class TestPregnancySettings(unittest.TestCase):
+    def setUp(self):
+        load_clan_settings()
+        self.pregnancy_class = Pregnancy_Events()
+        Cat.disable_random = True
+
+    def test_single_parent(self):
+        parent1 = Cat(moons=50, gender="male")
+        parent2 = Cat(moons=50, gender="female")
+
+        # our parents are unmated, thus this would be single parenthood
+        # allowed
+        self.assertTrue(
+            self.pregnancy_class.check_if_can_have_kits(
+                parent1,
+                allow_single_parent=True,
+                allow_unmated=True,
+                allow_affair=True,
+            )
+        )
+        # not allowed
+        self.assertFalse(
+            self.pregnancy_class.check_if_can_have_kits(
+                parent1,
+                allow_single_parent=False,
+                allow_unmated=True,
+                allow_affair=True,
+            )
+        )
+
+        # set mate
+        parent1.mate = [parent2.ID]
+        # single parentage setting shouldn't prevent these cats from having kits
+        self.assertTrue(
+            self.pregnancy_class.check_if_can_have_kits(
+                parent1,
+                allow_single_parent=True,
+                allow_unmated=True,
+                allow_affair=True,
+            )
+        )
+        self.assertTrue(
+            self.pregnancy_class.check_if_can_have_kits(
+                parent1,
+                allow_single_parent=False,
+                allow_unmated=True,
+                allow_affair=True,
+            )
+        )
+
+    def test_unmated(self):
+        parent1 = Cat(moons=50, gender="male")
+        parent2 = Cat(moons=50, gender="female")
+
+        # parent1 is unmated
+        # allow
+        self.assertTrue(
+            self.pregnancy_class.check_if_can_have_kits(
+                parent1,
+                allow_single_parent=True,
+                allow_unmated=True,
+                allow_affair=True,
+            )
+        )
+        # don't allow
+        set_clan_setting("unmated parentage", False)
+        self.assertFalse(
+            self.pregnancy_class.check_if_can_have_kits(
+                parent1,
+                allow_single_parent=True,
+                allow_unmated=False,
+                allow_affair=True,
+            )
+        )
+
+        # set mate
+        parent1.mate = [parent2.ID]
+        # unmated parentage setting shouldn't prevent these cats from having kits
+        self.assertTrue(
+            self.pregnancy_class.check_if_can_have_kits(
+                parent1,
+                allow_single_parent=True,
+                allow_unmated=True,
+                allow_affair=True,
+            )
+        )
+        self.assertTrue(
+            self.pregnancy_class.check_if_can_have_kits(
+                parent1,
+                allow_single_parent=True,
+                allow_unmated=False,
+                allow_affair=True,
+            )
+        )
+
+    def test_affair(self):
+        parent1 = Cat(moons=50, gender="male")
+        parent2 = Cat(moons=50, gender="female")
+
+        # our parents are unmated and so an affair isn't allowed (only mated cats have affairs)
+        # allowed
+        self.assertFalse(
+            self.pregnancy_class.check_if_can_have_kits(
+                parent1,
+                allow_single_parent=False,
+                allow_unmated=False,
+                allow_affair=True,
+            )
+        )
+        # set mate
+        parent1.mate = [parent2.ID]
+        affair_cat = Cat(moons=50, gender="female")
+        parent1.relationships[affair_cat.ID] = Relationship(
+            cat_from=parent1, cat_to=affair_cat, romance=100
+        )
+        # our parents ARE mated and so an affair is allowed
+        # TODO: would love to test this, but the way affairs are set up does not allow it
+        # self.assertEqual(
+        #    self.pregnancy_class.get_second_parent(parent1), (affair_cat, True)
+        # )
+
+        # turn affairs off
+        set_clan_setting("affair", False)
+        # now an affair isn't allowed
+        self.assertEqual(
+            self.pregnancy_class.get_second_parent(parent1), (parent2, False)
+        )
+
+
 class CanHaveKits(unittest.TestCase):
+    def test_other_clan_pregnancy(self):
+        test_clan = Clan(name="clan")
+        test_clan.pregnancy_data = {}
+        cat = Cat(disable_random=True, moons=50)
+        game.used_group_IDs["5"] = CatGroup.OTHER_CLAN
+        cat.status.add_to_group("5")
+
+        self.assertFalse(Pregnancy_Events.handle_having_kits(cat, test_clan))
+
+    def test_cat_not_working_pregnancy(self):
+        test_clan = Clan(name="clan")
+        test_clan.pregnancy_data = {}
+        cat = Cat(disable_random=True, moons=50)
+        cat.get_injured("broken bone")
+
+        self.assertFalse(Pregnancy_Events.handle_having_kits(cat, test_clan))
+
     def test_prevent_kits(self):
         # given
         cat = Cat(disable_random=True)
